Adding stability enhancements
