+++ trcflow/flumen/process.go
@@ -223,6 +223,9 @@ func ProcessFlows(pluginConfig map[string]interface{}, logger *log.Logger) error
 	tfmContext.ExtensionAuthData, err = trcvutils.GetJSONFromClientByPost(config, httpClient, extensionAuthComponents["authHeaders"].(map[string]string), extensionAuthComponents["authUrl"].(string), extensionAuthComponents["bodyData"].(io.Reader))
 	if err != nil {
 		eUtils.LogErrorObject(config, err, false)
+		tfmContext.ExtensionAuthDataReloader = make(map[string]interface{}, 1)
+		tfmContext.ExtensionAuthDataReloader["config"] = config
+		tfmContext.ExtensionAuthDataReloader["identityConfig"] = trcIdentityConfig
 		//return err
 	}
